test: extract NetworkTestSupport (MockURLProtocol + makeMockURLSession) from TestSupport (#560, #434)
Seventh slice of splitting the large TestSupport.swift into domain
files. Slices 434a-f shipped through #565.

Moves the network fake — MockURLProtocol (a URLProtocol subclass that
intercepts URLSession requests) and its factory helper makeMockURLSession()
(which builds a URLSession with configuration.protocolClasses =
[MockURLProtocol.self]) — byte-identical from TestSupport.swift lines
450-483 into a new NetworkTestSupport.swift with only `import
Foundation` (no @testable BugNarrator — the network fake never touches
internal product types; no XCTest).

The load-bearing line

    nonisolated(unsafe) static var requestHandler:
        ((URLRequest) throws -> (HTTPURLResponse, Data))?

is preserved verbatim; the cross-test-suite handshake continues to work
(four suites — IssueExtractionServiceTests, TranscriptionClientTests,
GitHubExportProviderTests, JiraExportProviderTests — set it to inject
responses).

Byte-diff against origin/main lines 450-483 is silent (exit 0) —
byte-identity proven. 662 unit tests pass unchanged; docs-sync
validation passes.

Pre-build spec reviewed by codex in a fresh session; 3 findings folded:
also move makeMockURLSession() (adjacent + same 4 users), Foundation-only
imports, byte-diff acceptance with exact range 450-483 (I initially
miscounted the closing brace of makeMockURLSession as line 482; codex
correctly gave 479-483 for that helper — a one-line under-extraction
first surfaced as a compile error and was fixed before commit).
