+++ Tests/BugNarratorTests/TestSupport.swift
@@ -447,41 +447,6 @@ func waitUntil(
     }
 }
 
-final class MockURLProtocol: URLProtocol {
-    nonisolated(unsafe) static var requestHandler: ((URLRequest) throws -> (HTTPURLResponse, Data))?
-
-    override class func canInit(with request: URLRequest) -> Bool {
-        true
-    }
-
-    override class func canonicalRequest(for request: URLRequest) -> URLRequest {
-        request
-    }
-
-    override func startLoading() {
-        guard let requestHandler = Self.requestHandler else {
-            fatalError("MockURLProtocol.requestHandler was not set.")
-        }
-
-        do {
-            let (response, data) = try requestHandler(request)
-            client?.urlProtocol(self, didReceive: response, cacheStoragePolicy: .notAllowed)
-            client?.urlProtocol(self, didLoad: data)
-            client?.urlProtocolDidFinishLoading(self)
-        } catch {
-            client?.urlProtocol(self, didFailWithError: error)
-        }
-    }
-
-    override func stopLoading() {}
-}
-
-func makeMockURLSession() -> URLSession {
-    let configuration = URLSessionConfiguration.ephemeral
-    configuration.protocolClasses = [MockURLProtocol.self]
-    return URLSession(configuration: configuration)
-}
-
 /// In-memory `ExportReceiptStoring` for tests that need to drive the receipt
 /// lifecycle — e.g. seeding a `.pending` receipt to exercise the reconcile
 /// path — without the file-backed production store. Records `markSucceeded`
